+++ src/Client.php
@@ -27,7 +27,7 @@ use Qless\Workers\Collection as WorkersCollection;
  * @method int unrecur(string $jid)
  * @method bool|string fail(string $jid, string $worker, string $group, string $message, ?string $data = null)
  * @method string[] jobs(string $state, int $offset = 0, int $count = 25)
- * @method bool|string get(string $jid)
+ * @method false|string get(string $jid)
  * @method string multiget(string[] $jid)
  * @method string complete(string $jid, string $workerName, string $queueName, string $data, ...$args)
  * @method void timeout(string $jid)
