+++ src/Jobs/RecurringJob.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Qless\Jobs;
+
+/**
+ * Qless\Jobs\RecurringJob
+ *
+ * @package Qless\Jobs
+ */
+class RecurringJob extends BaseJob
+{
+}
