+++ src/Jobs/Reservers/OrderedReserver.php
@@ -2,7 +2,7 @@
 
 namespace  Qless\Jobs\Reservers;
 
-use Qless\Jobs\Job;
+use Qless\Jobs\BaseJob;
 
 /**
  * Qless\Jobs\Reservers\OrderedReserver
@@ -16,12 +16,12 @@ class OrderedReserver extends AbstractReserver implements ReserverInterface
     /**
      * {@inheritdoc}
      *
-     * @return Job|null
+     * @return BaseJob|null
      */
-    final public function reserve(): ?Job
+    final public function reserve(): ?BaseJob
     {
         foreach ($this->queues as $queue) {
-            /** @var \Qless\Jobs\Job|null $job */
+            /** @var \Qless\Jobs\BaseJob|null $job */
             $job = $queue->pop($this->worker);
             if ($job !== null) {
                 return $job;
