+++ tests/Jobs/Reservers/OrderedReserverTest.php
@@ -2,7 +2,7 @@
 
 namespace Qless\Tests\Jobs\Reservers;
 
-use Qless\Jobs\Job;
+use Qless\Jobs\BaseJob;
 use Qless\Jobs\Reservers\OrderedReserver;
 use Qless\Queues\Queue;
 use Qless\Tests\QlessTestCase;
@@ -32,7 +32,7 @@ class OrderedReserverTest extends QlessTestCase
         $_SERVER['performed'] = '';
 
         $class = new class {
-            public function perform(Job $job): void
+            public function perform(BaseJob $job): void
             {
                 $_SERVER['performed'] = "{$job->queue}:{$job->data[0]}";
             }
